+++ features/config/perennial_branches/display.feature
@@ -2,7 +2,7 @@ Feature: display the perennial branches
 
   Scenario: unconfigured
     Given the perennial branches are not configured
-    When I run "git-town perennial-branches"
+    When I run "git-town config perennial-branches"
     Then it prints:
       """
       [none]
@@ -10,7 +10,7 @@ Feature: display the perennial branches
 
   Scenario: configured
     Given the perennial branches are "qa" and "production"
-    When I run "git-town perennial-branches"
+    When I run "git-town config perennial-branches"
     Then it prints:
       """
       qa
