+++ features/config/perennial_branches/update.feature
@@ -6,14 +6,14 @@ Feature: change the perennial branches
 
   Scenario: add a perennial branch
     Given the perennial branches are "qa"
-    When I run "git-town perennial-branches update" and answer the prompts:
+    When I run "git-town config perennial-branches update" and answer the prompts:
       | PROMPT                            | ANSWER               |
       | Please specify perennial branches | [DOWN][SPACE][ENTER] |
     Then the perennial branches are now "qa" and "staging"
 
   Scenario: remove a perennial branch
     Given the perennial branches are "staging" and "qa"
-    When I run "git-town perennial-branches update" and answer the prompts:
+    When I run "git-town config perennial-branches update" and answer the prompts:
       | PROMPT                            | ANSWER               |
       | Please specify perennial branches | [DOWN][SPACE][ENTER] |
     Then the perennial branches are now "qa"
