+++ features/shared/run_outside_repo.feature
@@ -9,14 +9,14 @@ Feature: require a Git repository
       """
 
     Examples:
-      | COMMAND                     |
-      | git-town config             |
-      | git-town config main-branch |
-      | git-town perennial-branches |
-      | git-town hack feature       |
-      | git-town kill               |
-      | git-town new-pull-request   |
-      | git-town prune-branches     |
-      | git-town repo               |
-      | git-town ship               |
-      | git-town sync               |
+      | COMMAND                            |
+      | git-town config                    |
+      | git-town config main-branch        |
+      | git-town config perennial-branches |
+      | git-town hack feature              |
+      | git-town kill                      |
+      | git-town new-pull-request          |
+      | git-town prune-branches            |
+      | git-town repo                      |
+      | git-town ship                      |
+      | git-town sync                      |
