+++ features/shared/wrong_parameter_count.feature
@@ -19,12 +19,12 @@ Feature: too few or many parameters
       | config main-branch arg1 arg2          |
       | config new-branch-push-flag arg1 arg2 |
       | config offline arg1 arg2              |
+      | config perennial-branches arg1        |
+      | config perennial-branches update arg1 |
       | hack                                  |
       | hack arg1 arg2                        |
       | kill arg1 arg2                        |
       | new-pull-request arg1                 |
-      | perennial-branches arg1               |
-      | perennial-branches update arg1        |
       | prepend                               |
       | prune-branches arg1                   |
       | pull-branch-strategy arg1 arg2        |
